backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
requests==2.31.0
google-api-python-client==2.122.0
google-auth==2.28.2
